module_14/contacts_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
fastapi-mail==1.4.1
cloudinary==1.41.0
slowapi==0.1.9
redis==5.2.0
aiofiles==24.1.0
Jinja2==3.1.4
